backend/requirements.txt
# StarGuide AI Mentor - Backend Dependencies
# FastAPI & Web Framework
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-multipart>=0.0.9

# Database & Caching
motor==3.3.1
pymongo==4.5.0
redis>=5.0.0
beanie>=1.24.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pyjwt>=2.10.1

# AI Integrations
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.4.0

# Real-time & WebSocket
python-socketio>=5.11.0
eventlet>=0.33.3

# Utilities & Environment
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
aiofiles>=23.2.1
httpx>=0.27.0

# Monitoring & Analytics
structlog>=24.1.0
prometheus-client>=0.19.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.1
isort>=5.13.2
mypy>=1.8.0